requirements.txt
apypie
pytest-base-url
pytest-html
pytest-selenium
pytest-testinfra
pytest-variables
pytest-xdist
